+++ pkg/kv/kvserver/kvserverbase/base.go
@@ -205,3 +205,11 @@ func IntersectSpan(
 	}
 	return
 }
+
+// SplitByLoadMergeDelay wraps "kv.range_split.by_load_merge_delay".
+var SplitByLoadMergeDelay = settings.RegisterDurationSetting(
+	"kv.range_split.by_load_merge_delay",
+	"the delay that range splits created due to load will wait before considering being merged away",
+	5*time.Minute,
+	settings.NonNegativeDuration,
+)
